# src/main/resources/static/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "TGM Schulplaner API",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0"
    },
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/v1/group/{gid}": {
      "get": {
        "tags": [
          "group"
        ],
        "summary": "Get a group by its id if the authenticated user has access",
        "operationId": "getGroup",
        "parameters": [
          {
            "name": "gid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Group"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer-key": [

            ]
          }
        ]
      },
      "put": {
        "tags": [
          "group"
        ],
        "summary": "Modify a group by id",
        "operationId": "modifyGroup",
        "parameters": [
          {
            "name": "gid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModifyGroup"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Group"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer-key": [

            ]
          }
        ]
      }
    },
    "/api/v1/group": {
      "post": {
        "tags": [
          "group"
        ],
        "summary": "Create a group",
        "operationId": "createGroup",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewGroup"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Group"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer-key": [

            ]
          }
        ]
      }
    },
    "/api/v1/group/{gid}/member/{uid}": {
      "put": {
        "tags": [
          "group"
        ],
        "summary": "Modify a group member for a group",
        "operationId": "modifyMember",
        "parameters": [
          {
            "name": "gid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModifyMember"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "security": [
          {
            "bearer-key": [

            ]
          }
        ]
      }
    },
    "/api/v1/group/{gid}/member": {
      "post": {
        "tags": [
          "group"
        ],
        "summary": "Modify a group member for a group",
        "operationId": "createMember",
        "parameters": [
          {
            "name": "gid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewMember"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "security": [
          {
            "bearer-key": [

            ]
          }
        ]
      },
      "delete": {
        "tags": [
          "group"
        ],
        "summary": "Modify a group by id",
        "operationId": "modifyMember_1",
        "parameters": [
          {
            "name": "gid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "user",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "security": [
          {
            "bearer-key": [

            ]
          }
        ]
      }
    },
    "/api/v1/groups": {
      "get": {
        "tags": [
          "group"
        ],
        "summary": "Get all groups the authenticated user is a member of",
        "operationId": "getGroups",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Group"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer-key": [

            ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Group": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "NewGroup": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "ModifyMember": {
        "type": "object",
        "properties": {
          "admin": {
            "type": "boolean"
          }
        }
      },
      "NewMember": {
        "type": "object",
        "properties": {
          "uid": {
            "type": "string",
            "format": "uuid"
          },
          "admin": {
            "type": "boolean"
          }
        }
      },
      "ModifyGroup": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "settings": {
            "type": "object",
            "additionalProperties": {
              "type": "string",
              "enum": [
                "READ",
                "CREATE",
                "MODIFY",
                "DELETE"
              ]
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer-key": {
        "type": "http",
        "name": "bearer-key",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "jwt"
      }
    }
  }
}